# [WIP] Add arenaskl optimization for boundary checks