Update arenaskl to not check `{lower,upper}bound` in `Prev` and `Next`. Instead, we compare pointers to the nodes which bound the lower and upper bounds. This is described in #98.

The benchmarks show a very significant increase in performance:
```
name             old time/op  new time/op  delta
IterBoundNext    19.8ns ± 0%  12.7ns ± 0%  -35.99%  (p=0.000 n=10+7)
IterBoundNext-8  19.9ns ± 1%  12.7ns ± 0%  -36.28%  (p=0.000 n=10+8)
IterBoundPrev    19.9ns ± 1%  12.8ns ± 1%  -35.86%  (p=0.000 n=10+10)
IterBoundPrev-8  19.9ns ± 1%  12.8ns ± 1%  -35.78%  (p=0.000 n=10+10)
```

However, this performance gain does not reflect real life performance gains. In the benchmark, we use the maximum possible key as the upper bound, and the minimum possible key as the lower bound. 

To see what real life performance gains would look like, we would need to know what the average number of iterations is before hitting the `{lower,upper}bound`. I suspect that if the `{lower,upper}bound` is reached after only a few iterations, the cost of the `SeekGE(upper)` and `SeekLT(lower)` would overcome the cost of doing a direct boundary comparison a few times.

I don't really have a ballpark on what the average number of iterations in CRDB is. Do you have any insights?